[REFACTOR] (Input) : input class 통합
- 여러개의 함수로 나뉘어져있던 input 함수를 Input class로 통합
- input 이라는 폴더명을 src 바로 아래로 수정
(input class로 만들게 되면서 컴포넌트 개념과는 많이 달라졌다고 생각했기 때문)

여러개의 파일로 나뉘어져있던 input 함수들을 하나의 Input class로 통합하여 사용하였습니다
