+++ src/App.js
@@ -1,13 +1,12 @@
-import inputCarList from "./Components/Input/inputCarList.js";
-import inputTryNum from "./Components/Input/inputTryNum.js";
 import printWinner from "./Components/Output/printWinner.js";
 import raceCar from "./Components/RaceCar/raceCar.js";
+import Input from "./Input/Input.js";
 
 class App {
   async run() {
     try {
-      const carList = await inputCarList();
-      const tryNumber = await inputTryNum();
+      const carList = await Input.CarList();
+      const tryNumber = await Input.TryNum();
       const winnerCars = raceCar(carList, tryNumber);
       printWinner(winnerCars);
     } catch (error) {
