--- src/Components/Input/module/printGideMessage.js
@@ -1,7 +1,0 @@
-import { Console } from "@woowacourse/mission-utils";
-
-const printGideMessage = async (gideMessage) => {
-  const input = await Console.readLineAsync(gideMessage);
-  return String(input);
-};
-export default printGideMessage;
